requirements.txt
google-adk>=1.15.0
python-dotenv>=1.0.1
